refactor(libs-wifi): WiFi機能を独立ライブラリ(@libs-wifi)に分離
- apps/dashboard/src/app/wifi 配下のコードを libs/wifi に移行
- Nxライブラリ @libs-wifi を生成し、プロジェクト名を libs-wifi に設定
- WiFi関連コンポーネント、サービス、関数、モデルを移行
- wifi.model.ts と wifi.models.ts を統合して単一ファイル化
- tsconfig.base.json に @libs-wifi パスエイリアスを追加
- apps/dashboard/src/app/console/console.component.ts のインポートを更新
- libs/dashboard/models から WiFiモデルを削除
- libs/dashboard/utils から wifi.utils の re-export を削除
- ビルド・テスト確認済み
