--- apps/dashboard/src/app/wifi/functions/index.ts
@@ -1,5 +1,0 @@
-export {
-  parseWifiIfconfigOutput,
-  parseWifiIwconfigOutput,
-  parseWifiIwlistOutput,
-} from './wifi-parser';
