--- apps/dashboard/src/app/wifi/services/index.ts
@@ -1,5 +1,0 @@
-// 存在しないパスへの参照を削除
-// TODO: 必要に応じて正しいパスに修正
-// export * from '../../wifi-setting-containers/services/command-line.service';
-// export * from '../../wifi-setting-containers/services/monaco-editor.service';
-// export * from '../../wifi-setting-containers/services/web-serial.service';
