+++ libs/dashboard/utils/src/index.ts
@@ -3,4 +3,3 @@ export * from './lib/date.utils';
 export * from './lib/errors';
 export * from './lib/file.utils';
 export * from './lib/parser.utils';
-export * from './lib/wifi.utils';
