+++ libs/wifi/README.md
@@ -1,0 +1,7 @@
+# wifi
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test wifi` to execute the unit tests.
