+++ libs/wifi/ng-package.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/libs/wifi",
+  "lib": {
+    "entryFile": "src/index.ts"
+  }
+}
