+++ libs/wifi/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@libs-wifi",
+  "version": "0.0.1",
+  "peerDependencies": {
+    "@angular/common": "^20.3.0",
+    "@angular/core": "^20.3.0"
+  },
+  "sideEffects": false
+}
