+++ libs/wifi/src/index.ts
@@ -1,0 +1,9 @@
+export * from './lib/components/wifi-info/wifi-info.component';
+export * from './lib/components/wifi/wifi.component';
+export {
+  parseWifiIfconfigOutput,
+  parseWifiIwconfigOutput,
+  parseWifiIwlistOutput,
+} from './lib/functions/wifi-parser';
+export * from './lib/models/wifi.model';
+export * from './lib/services/wifi.service';
