+++ libs/wifi/src/lib/components/wifi-info/wifi-info.component.spec.ts
@@ -1,4 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { WiFiInformation } from '../../models/wifi.model';
 import { WifiInfoComponent } from './wifi-info.component';
 
 describe('WifiInfoComponent', () => {
@@ -12,6 +13,14 @@ describe('WifiInfoComponent', () => {
 
     fixture = TestBed.createComponent(WifiInfoComponent);
     component = fixture.componentInstance;
+    component.wifiInfo = {
+      SSID: 'test-ssid',
+      address: '00:00:00:00:00:00',
+      channel: 1,
+      frequency: '2.4 GHz',
+      quality: 'Quality=70/70',
+      spec: 'WPA2',
+    } as WiFiInformation;
     fixture.detectChanges();
   });
