+++ libs/wifi/src/lib/components/wifi-info/wifi-info.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
-import { WiFiInformation } from '../../models/wifi.models';
+import { WiFiInformation } from '../../models/wifi.model';
 
 @Component({
   selector: 'choh-wifi-info',
