+++ libs/wifi/src/lib/components/wifi/wifi.component.ts
@@ -5,8 +5,8 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { DialogService } from '@libs-dialogs';
 import { ButtonComponent } from '@libs-ui';
-import { dummyWiFiInformation } from '@dashboard/models';
-import { WifiInfoComponent } from './components/wifi-info/wifi-info.component';
+import { dummyWiFiInformation } from '../../models/wifi.model';
+import { WifiInfoComponent } from '../wifi-info/wifi-info.component';
 
 @Component({
   selector: 'choh-wifi',
