+++ libs/wifi/src/lib/functions/wifi-parser.ts
